==> 01_generating_data.ipynb <==
# Advanced Big Data Application

2.	Download the latest version from: https://www.kaggle.com/Cornell-University/arxiv  (This is a large json file ~2.81 GB). To deal with json files in python, please refer to the working JSON tutorial. 

3.	Create a new dataset in csv format with only records containing comments and without the word COVID in the title. This new dataset will have only 5 columns: id, title, comments, journal-ref and categories. 

# Module Imports
import dask.dataframe as dd
from dask.distributed import Client, LocalCluster

# Module Imports
import dask.dataframe as dd
from dask.distributed import Client, LocalCluster

# Sets up a dask cluter to run on local machine
cluster = LocalCluster() 

# Sets the maximum number of workers 
cluster.adapt(minimum=1, maximum=8) 

# Creates a client to enable interaction with the cluster
client = Client(cluster)

# Explicitly defining data types
defined_dtypes = {'id': str, 'title': str, 'comments':  str, 'journal-ref':  str, 'id':  str}
# reads in data
df = dd.read_json('arxiv.json', blocksize="256 MiB", dtype=defined_dtypes, )
df.head(5)

defined_dtypes = {'id': str, 'title': str, 'comments':  str, 'journal-ref':  str, 'id':  str}

df = dd.read_json('arxiv.json', blocksize="256 MiB", dtype=defined_dtypes, )
df.head(5)


## Removing papers without comments

df = df[['id', 'title', 'comments', 'journal-ref', 'categories']]

df = df[df['comments']!='None']
df.head(5)

# Filter to remove records containing 'COVID' in the title 
df = df.loc[~df["title"].str.contains("covid", case=False)]

# Some titles included new line characters
df['title'] = df['title'].str.replace('\n', '').replace('"', '')

final_df = df.compute()

final_df.to_csv("without_covid.csv", index=False)

client.close()
