==> 02_cleaning.ipynb <==
# Module Imports
import dask.dataframe as dd
from dask.distributed import Client, LocalCluster
import matplotlib.pyplot as plt
import seaborn as sns
from util import clean_journal_ref

cluster = LocalCluster() 
cluster.scale(8) 

# Sets the number of workers 
cluster.adapt(minimum=1, maximum=8) 

# Allows the cluster to auto scale to 10 when tasks are computed 
client = Client(cluster)

defined_dtypes = {'id': str, 'title': str, 'comments':  str, 'journal-ref':  str, 'journal-ref_cleaned':  str,
                  'categories':  str, 'categories_original':str, "journal-ref_original":str,
                  'feild':str, 'category':str, 'sub_category':str, 'first_category':str}

df = dd.read_csv('without_covid.csv', blocksize="256 MiB", dtype=defined_dtypes)
df.head(5)

for col in df.columns:
    # make all text entries lower case
    df[col] = df[col].str.lower()
    
    # remove new line characters
    df[col] = df[col].str.replace('\n', '')
    
    # strip leading and trailing spaces
    df[col] = df[col].str.strip()

df.head(5)

df['journal-ref_cleaned'] = df['journal-ref'].apply(clean_journal_ref, meta=('journal-ref', str))

# Removing rows where comments are logged as NaN
df = df.dropna(subset=['comments'])  
# Removing rows where comments are logged as None
df = df[df['comments']!='None']

# Select only the first category mentioned
df['first_category'] = df['categories'].str.split(' ').str[0]
print(type(df['first_category'][0]))
df.head(5)

df[['category', 'sub_category']] = df['categories'].str.split(".", expand=True, n=2)

df['sub_category'] = df['sub_category'].str.split(' ').str[0]
df.head(5)

df.head(15)

field_dict = {'astro-ph':'physics',
             'cond-mat':'physics',
             'gr-qc':'physics',
             'hep-ex':'physics',
             'hep-lat':'physics',
             'hep-ph':'physics',
             'hep-th':'physics',
             'math-ph':'physics',
             'nlin':'physics',
             'nucl-ex':'physics',
             'nucl-th':'physics',
             'physics':'physics',
             'quant-ph':'physics',
              'math':'math',
              'CoRR':'computer science',
              'q-bio':'quantative biology',
              'q-fin':'quantative finance',
              'stat':'statistics',
              'eess':'electrical engineering and systems science',
              'econ':'economics'}

df['feild'] = df['category'].map(field_dict)

df = df.rename(columns={"categories": "categories_original", "journal-ref": "journal-ref_original"}) 

final_df = df.compute()

final_df.head(5)

final_df.to_csv("without_covid_cleaned.csv", index=False)

cluster.close()

def clean_journal_ref(journal):
    # Select only strings to this cleaning is not applied to Nan values
    if type(journal) == str:
        # Errtum shows changes and is usually followed by the journal name 
        # again and a corrected page number or edition number
        # Split string on 'erratum' and remove second half
        journal = journal.split("erratum")[0]

        # make all strings start and end with a dot (.0)
        journal = '.' + journal + '.'

        # replace any spacial character or space with a .
        journal = re.sub('[\W_]', '.', journal)

        # Remove numbers
        journal = re.sub('\d+', '.', journal)

        words_to_replace = ['pp', 'no', 'volume', 'vol', 'issue', 'pages',]
        # Remove month names
        month_names = list(month_name[1:])
        month_names_short = [x[3:] for x in month_names]

        words_to_replace += month_names
        words_to_replace += month_names_short


        for mini_word in words_to_replace:
            journal = journal.replace('.'+mini_word+'.', '.')

        # replace common abbreviations
        for key, value in {'.j.': '.journal.of.', '.rev.': '.review.',
                           '.phys.': '.physics.', '.physical.': '.physics.'}.items():
            journal = journal.replace(key, value)

        # Remove single letters
        journal = re.sub('[.][a-zA-Z](?:[.]|$)', '.', journal)

        # remove repeated dots
        journal = re.sub('\.{2,}', '.', journal)

        # strip dots at the start and end
        journal = journal.strip('.')

    return journal
