==> 04_get_word_rankings.ipynb <==
# Module Imports
import dask.dataframe as dd
from dask.distributed import Client, LocalCluster
import matplotlib.pyplot as plt
import seaborn as sns
import pandas as pd

cluster = LocalCluster() 
cluster.scale(8) 

# Sets the number of workers 
cluster.adapt(minimum=1, maximum=8) 

# Allows the cluster to auto scale to 10 when tasks are computed 
client = Client(cluster)

defined_dtypes = {'id': str, 'title': str, 'comments':  str, 'journal-ref':  str, 'id':  str}

df = dd.read_csv('without_covid_cleaned.csv', blocksize="256 MiB", dtype=defined_dtypes)
df.head(5)

titles = df['title'].compute()

titles = titles.str.lower()

titles = titles.str.replace('[^A-Za-z\s]', '', regex=True)
titles = titles.str.replace('  ', ' ', regex=True)

for i in range(5):
    print(titles[i])

from nltk.corpus import stopwords

# get enlish stop words
english_stopwords = stopwords.words('english')

# remove any non alphabetic, non whitespace characters
my_english_stopwords = [word.replace('[^A-Za-z\s]', '') for word in english_stopwords]

my_english_stopwords[:10]

# Make a list with each word as an item
titles_as_lists = titles.str.split()
# Make each word its own item in a series
individual_words = titles_as_lists.explode()

print(individual_words.head(10))

import nltk
from nltk.stem import PorterStemmer

stemmer = PorterStemmer()

# Applying lemmatisation to series
individual_word_stems = individual_words.apply(lambda word: stemmer.stem(word))

individual_word_stems.head(10)

import nltk
from nltk.stem import PorterStemmer

stemmer = PorterStemmer()

for i in ['model', 'modelling', 'modelled', 'models']:
    print(i, ':', stemmer.stem(i))

individual_words.value_counts().head(10)

# Caluclate frequency of each word in series
value_counts = individual_word_stems.value_counts()

**Word Rankings before stop words have been removed**
- of      899003
- the     689110
- and     540041
- in      537339
- a       346181
- for     337998
- on      217102
- with    196198
- to      121488
- from     98387

value_counts_no_stopwords = value_counts.drop(index=my_english_stopwords, errors='ignore')
value_counts_no_stopwords.head(10)

Before lemmitisation

quantum     98155

model       74674

theory      52267

using       50263

systems     46438

learning    42634

models      39951

field       39216

analysis    34508

magnetic    33020


final_df = pd.DataFrame(value_counts_no_stopwords)

final_df['rank'] = final_df['count'].rank(ascending=False)
final_df.head(10)

final_df.to_csv("word_rankings.csv", index=True)

client.close()
