==> 06_make_reccomendations.ipynb <==
Based on the ranking of words, pick the top ten and recommend a journal for each word. 

import pandas as pd
import dask.dataframe as dd
from dask.distributed import Client, LocalCluster
from util import clean_journal_ref

word_rankings = pd.read_csv("word_rankings.csv")
top_10_words = word_rankings['title'][:10]
top_10_words

cluster = LocalCluster() 
cluster.scale(8) 

# Sets the number of workers 
cluster.adapt(minimum=1, maximum=8) 

# Allows the cluster to auto scale to 10 when tasks are computed 
client = Client(cluster)

defined_dtypes = {'id': str, 'title': str, 'comments':  str, 'journal-ref':  str, 'journal-ref_cleaned':  str,
                  'categories':  str, 'categories_original':str, "journal-ref_original":str,
                  'feild':str, 'category':str, 'sub_category':str}

df = dd.read_csv('without_covid_cleaned.csv', blocksize="256 MiB", dtype=defined_dtypes)
df.head(5)

journal_totals = df['journal-ref_cleaned'].value_counts().compute()
journal_totals_dict = journal_totals.to_dict()
print(journal_totals_dict)

import nltk
from nltk.stem import PorterStemmer

stemmer = PorterStemmer()

def lemmatize_sentence(sentence):
    # initial lemmatiser
    stemmer = PorterStemmer()
    # Tokenize the sentence into individual words and apply lemmatiser
    lemmatized_words = [stemmer.stem(word) for word in nltk.word_tokenize(sentence)]
    # return words reconnected as a sentence
    return ' '.join(lemmatized_words)

# First we have to apply the same lemmatisation to the titles as we did to the key words
titles_lemmatised = df['title'].apply(lemmatize_sentence, meta=('title', str)).compute()

def reccommend_journals(keyword):    
    #select only rows where title contains keyword
    keyword_df = df[titles_lemmatised.str.contains(keyword)]

    #group by journal-ref
    journal_frequency = keyword_df.groupby('journal-ref_cleaned').size()

    return journal_frequency

writer = pd.ExcelWriter('journal_reccomendations.xlsx')

for keyword in top_10_words:
    # apply reccommend_journals to return the no. articles containing the keyword for each journal-ref
    top_journals = reccommend_journals(keyword).compute()

    # Create an empty dataframe to populate
    top_journals_df = pd.DataFrame({'top_journals':top_journals.index, 'count':top_journals.values})

    # create a total column for the number of articles containing the keyword
    top_journals_df['total']= top_journals_df['top_journals'].map(journal_totals_dict)
    
    # create a percentage column that divides total by the total no. articles in that journal-ref
    top_journals_df['perc']= (top_journals_df['count']/top_journals_df['total'])*100

    # only return journals with more than 100 articles.
    valid_journals = top_journals_df[top_journals_df['total']>=100]
    
    # save the resulting datafram to a sheet with the keyword as the sheetname
    valid_journals.to_excel(writer, sheet_name=keyword, index=False)
    
    print(keyword, 'done :)')

#client.close()

writer.close()
